Aloisio magalhaes updates (#10)
* Refactor .gitignore file

Remove unnecessary entries and add new ones for improved project management and build process

Relates to #123

* Adds TypeScript support and automatic build mode to CodeQL workflow

This commit modifies the CodeQL workflow to include TypeScript as a supported language and automatic build mode. This change enables the analysis of code written in TypeScript and ensures that the workflow automatically builds the project. By expanding language support and automating the build process, this commit improves the overall functionality and efficiency of the CodeQL workflow.

Fixes #N/A

* Updates build mode to 'autobuild' in codeql.yml

Changes the build mode from 'automatic' to 'autobuild' in the codeql.yml file. This modification ensures that the build process is automatically triggered when running CodeQL analysis.

* Updates build mode to manual in codeql.yml

Changes the build mode from autobuild to manual in the codeql.yml file. This allows for more control over the build process and ensures that the code is only built when explicitly triggered.

* Updates build mode to 'none' in CodeQL workflow

Removes the 'manual' build mode and sets it to 'none' in the CodeQL workflow. This change ensures that the workflow does not attempt to build the code.

Improves the efficiency of the workflow by skipping the unnecessary build step.

* fix: :bug: actions

error typescript

---------

Signed-off-by: AloisioMagalhães <educacao1ouro@gmail.com>